fix conversion algorythm for only cents Signed-off-by: tomekpapke <tomasz.papke@wp.pl>

+++ CurrencyConverter.Service/CurrencyWordValueComposer.cs
@@ -95,7 +95,7 @@ namespace CurrencyConverter.Service
 
         public CurrencyWordValueComposer ComposeIntegerExpression()
         {
-            if (currency == 0)
+            if ((int)currency == 0)
             {
                 sb.Append(ZERO).Append(" ");
                 return this;
